+++ src/main/java/wad/seoul_nolgoat/web/store/StoreController.java
@@ -8,12 +8,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wad.seoul_nolgoat.service.bookmark.BookmarkService;
 import wad.seoul_nolgoat.service.store.StoreService;
 import wad.seoul_nolgoat.web.store.dto.response.StoreDetailsDto;
-import wad.seoul_nolgoat.web.store.dto.response.StoreForBookmarkDto;
-
-import java.util.List;
 
 @Tag(name = "가게")
 @RequiredArgsConstructor
@@ -22,19 +18,11 @@ import java.util.List;
 public class StoreController {
 
     private final StoreService storeService;
-    private final BookmarkService bookmarkService;
 
     @Operation(summary = "가게 단건 조회")
     @GetMapping("/{storeId}")
     public ResponseEntity<StoreDetailsDto> showStoreByStoreId(@PathVariable Long storeId) {
         return ResponseEntity
                 .ok(storeService.findStoreWithReviewsByStoreId(storeId));
-    }
-
-    @Operation(summary = "유저 ID를 통한 즐겨찾기 가게 목록 조회")
-    @GetMapping("/bookmarked/{userId}")
-    public ResponseEntity<List<StoreForBookmarkDto>> showBookmarkedStoresByUserId(@PathVariable Long userId) {
-        return ResponseEntity
-                .ok(bookmarkService.findBookmarkedStoresByUserId(userId));
     }
 }
